requirements/base.txt
# Bleeding edge Django
django==1.10.4

# Configuration
django-environ==0.4.1
django-secure==1.0.1


# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1
django-floppyforms==1.7.0

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.29.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.6.0
redis>=2.10.5


celery==4.0.2


# Your custom requirements go here
django-guardian==1.4.6
djangorestframework==3.5.3
django-mptt==0.8.7
rest-condition==1.0.2
drf-nested-routers==0.11.1
django-cors-headers==1.3.1
djangorestframework-jwt==1.9.0
openpyxl==2.4.1
django-filter==1.0.1

# django model states
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-fsm-log==1.3.0

# Markdown support for view documentation
markdown==2.6.7
